Rename details form store file
